Backend/requirements.txt
fastapi
uvicorn
python-dotenv
pyotp
influxdb-client
requests
apscheduler
numpy
scikit-learn
Pillow
nanoid